# Optimize ManagedError ctor and (WasmRoot|WasmExternalRoot).clear