--- a/src/mono/wasm/runtime/roots.ts
+++ b/src/mono/wasm/runtime/roots.ts
@@ -287,8 +287,17 @@ class WasmJsOwnedRoot<T extends MonoObject> implements WasmRoot<T> {
     }
 
     set(value: T): T {
-        const destinationAddress = this.__buffer.get_address(this.__index);
-        cwraps.mono_wasm_write_managed_pointer_unsafe(destinationAddress, <ManagedPointer>value);
+        if (value === <any>0) {
+            // Don't use .set since it involves a gc write barrier, and the write barrier
+            //  is unnecessary overhead for zeroing that shows up in profiles.
+            // The write barrier is only necessary when storing a managed pointer into
+            //  a root because doing that changes its reachability/liveness
+            const address32 = this.__buffer.get_address_32(this.__index);
+            Module.HEAPU32[address32] = 0;
+        } else {
+            const destinationAddress = this.__buffer.get_address(this.__index);
+            cwraps.mono_wasm_write_managed_pointer_unsafe(destinationAddress, <ManagedPointer>value);
+        }
         return value;
     }
 
@@ -381,7 +390,15 @@ class WasmExternalRoot<T extends MonoObject> implements WasmRoot<T> {
     }
 
     set(value: T): T {
-        cwraps.mono_wasm_write_managed_pointer_unsafe(this.__external_address, <ManagedPointer>value);
+        if (value === <any>0) {
+            // Don't use .set since it involves a gc write barrier, and the write barrier
+            //  is unnecessary overhead for zeroing that shows up in profiles.
+            // The write barrier is only necessary when storing a managed pointer into
+            //  a root because doing that changes its reachability/liveness
+            Module.HEAPU32[<any>this.__external_address >>> 2] = 0;
+        } else {
+            cwraps.mono_wasm_write_managed_pointer_unsafe(this.__external_address, <ManagedPointer>value);
+        }
         return value;
     }