`ManagedError`'s constructor unintentionally retrieves the value of `Error.stack`, which is very expensive. This PR fixes that by using a workaround of constructing a temporary `Error` instead (still expensive, but much less expensive) so that we can only fetch the stack on demand when necessary. A potential future optimization would be to only do this in Chrome, since `super.stack` apparently works in Firefox (maybe Safari, too?) but the complication is probably not worth it at this point.

This PR also optimizes the root `clear` and `set` methods to not use a write barrier when zeroing a root, because the write barrier is expensive and not really necessary. The downside is this could cause an object to live slightly too long in some cases, but we call `clear` a LOT for correctness so the cost has a significant impact on interop speed, at least according to browser-bench.

Before:
| measurement | time |
|-:|-:|
|             JSInterop, LegacyExportInt |     6.9038ms |
|                 JSInterop, JSExportInt |     3.5533ms |
|          JSInterop, LegacyExportString |    33.4786ms |
|              JSInterop, JSExportString |    31.2959ms |
|                 JSInterop, JSImportInt |     3.3081ms |
|              JSInterop, JSImportString |    40.0709ms |
|            JSInterop, JSImportManyArgs |   227.4000ms |
|                JSInterop, JSImportTask |   299.7647ms |
|            JSInterop, JSImportTaskFail |  1089.6000ms |
|                JSInterop, JSImportFail |   165.3438ms |

After:
| measurement | time |
|-:|-:|
|             JSInterop, LegacyExportInt |     6.1863ms |
|                 JSInterop, JSExportInt |     3.6297ms |
|          JSInterop, LegacyExportString |    30.1000ms |
|              JSInterop, JSExportString |    28.9896ms |
|                 JSInterop, JSImportInt |     3.6116ms |
|              JSInterop, JSImportString |    38.5294ms |
|            JSInterop, JSImportManyArgs |   231.2609ms |
|                JSInterop, JSImportTask |   298.1176ms |
|            JSInterop, JSImportTaskFail |   809.1429ms |
|                JSInterop, JSImportFail |   159.5152ms |

JSImportTaskFail improved considerably due to the `ManagedError` stack optimization. The legacy exports meaningfully improved due to `clear` being optimized (part of the reason this matters is that right now to invoke a managed method, we HAVE to box the return value which means using a root that has to be cleared at the end.)

Note that these tests have significant timing variance from run to run because they spend a lot of time in the JS and C# garbage collectors (mostly due to the string allocations)