@kg 2022-10-27
Looks like using the class syntax to override `stack` isn't working, at least on CI. I'll have to try something else.