@pavelsavara 2022-10-27
I stolen your idea of inner Error instance and did my own attempt.
https://github.com/dotnet/runtime/pull/77547

I didn't try to benchmark it yet, will do that on Mon.