@kg 2022-10-27
Adding a test is a good idea, yours looks like it should also benchmark fine.