@kg 2022-10-28
This approach to optimizing ManagedError can't work because of a chrome bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1379185
Pavel's PR is probably the best we can do for that bit.