@pavelsavara 2022-11-01
I suggest that you revert the `ManagedError` change from this PR and we further explore the WasmRoot here.
Please also update the description and subject of the PR.