@kg 2022-11-05
Rebased and updated to remove the error stuff @pavelsavara